Adds ensemble import state management
Introduces a class to manage the state of ensemble imports,
including whether to create an esmry file, restart file patterns,
and the path to the parameter file. This allows for more robust
and flexible handling of ensemble data during import.

Also modifies summary case collection to use restart file patterns,
so that correct restart files are loaded for each case.