Updates ensemble curves and legend
Refactors ensemble curve updating to ensure correct legend ordering and visibility.

This change addresses an issue where the legend items were not consistently ordered due to the order curves were added to the plot.
It also fixes a bug where ensemble curves were still being generated even when hidden.

The update includes a new function to recreate the plot curve for the legend, and an optimization for hashing the realizations to avoid unnecessary re-creation of curves.