Adds capability configuration to field handles
Introduces a `configureCapabilities` virtual function to `PdmFieldHandle`to centralize the configuration of UI and XML capabilities for fields, avoiding code duplication.

This function is called during field initialization, allowing derived classes to customize capabilities based on specific requirements. A `setValueOptionsGenerator` is also added to `PdmUiFieldHandle` to allow setting the value options generator function.

This allows runtime configuration of value options for `PdmField` objects. For example, this is used to define acceptable font size values.