#12894 Stimplan model export: make sure MD is always increasing.
The numerical precision in the output was lowered in #11556 since Stimplan
was not correctly handling the high precision (requested by Nan).
That change leads to the MD not always increasing (which is expected by
Stimplan). Added another tweak to the MD values to make sure it increases
in the precision range that will be written to file.

The original issue asked for improved precision in the output,
but that is not a viable option this Stimplan does not use all
of the decimal precision.

Fixes #12894.