Creates refinement for radial grids with low cell count
Addresses an issue where radial grids with a low cell count in the J direction were not properly handled.

Now, if a radial grid has less than 4 cells in the J direction, the system creates a local grid refinement (LGR) to improve accuracy. Also adds a check if the grid is a full radial grid to create refinement for that case as well. This involves converting the radial grid to a Cartesian grid by transforming the node coordinates.

Refactors the LGR creation logic into reusable methods.